+++ History.md
@@ -1,5 +1,9 @@
 ## v.NEXT
 
+### Changes
+
+* The `meteor-babel` npm package has been updated to version 7.7.4.
+
 ## v1.8.2, 2019-11-14
 
 ### Breaking changes